move hardware challenges play to special folder
