+++ src/software/ai/hl/stp/play/hardware_challenge_plays/scoring_from_contested_possession_play.h
