+++ src/software/ai/hl/stp/play/hardware_challenge_plays/scoring_with_static_defenders_play.cpp
@@ -1,28 +1,29 @@
-#include "software/ai/hl/stp/play/pass_endurance_play.h"
+#include "software/ai/hl/stp/play/hardware_challenge_plays/scoring_with_static_defenders_play.h"
 
 #include "shared/constants.h"
 #include "software/ai/hl/stp/tactic/move/move_tactic.h"
 #include "software/util/generic_factory/generic_factory.h"
 
-PassEndurancePlay::PassEndurancePlay(std::shared_ptr<const PlayConfig> config)
+ScoringWithStaticDefendersPlay::ScoringWithStaticDefendersPlay(
+    std::shared_ptr<const PlayConfig> config)
     : Play(config, false)
 {
 }
 
-bool PassEndurancePlay::isApplicable(const World &world) const
+bool ScoringWithStaticDefendersPlay::isApplicable(const World &world) const
 {
     // This play is never applicable so it will never be chosen during gameplay
     // This play can be run for hardware challenges by using the Play override
     return false;
 }
 
-bool PassEndurancePlay::invariantHolds(const World &world) const
+bool ScoringWithStaticDefendersPlay::invariantHolds(const World &world) const
 {
     return false;
 }
 
-void PassEndurancePlay::getNextTactics(TacticCoroutine::push_type &yield,
-                                       const World &world)
+void ScoringWithStaticDefendersPlay::getNextTactics(TacticCoroutine::push_type &yield,
+                                                    const World &world)
 {
     std::vector<std::shared_ptr<MoveTactic>> move_tactics(NUM_ROBOTS);
     std::generate(move_tactics.begin(), move_tactics.end(),
@@ -31,10 +32,21 @@ void PassEndurancePlay::getNextTactics(TacticCoroutine::push_type &yield,
     do
     {
         TacticVector result = {};
-        if (world.gameState().isPlaying())
+        if (world.gameState().isStopped())
         {
-            // TODO (#2109): replace this example play with an actual implementation of
-            // pass endurance
+            // line up along center line
+            int initial_offset = static_cast<int>(-move_tactics.size() / 2 + 1);
+            for (size_t k = 0; k < move_tactics.size(); k++)
+            {
+                auto next_position = Point(
+                    world.field().centerPoint().x(),
+                    (initial_offset + static_cast<int>(k)) * 4 * ROBOT_MAX_RADIUS_METERS);
+                move_tactics[k]->updateControlParams(next_position, Angle::zero(), 0);
+            }
+        }
+        else if (world.gameState().isOurFreeKick())
+        {
+            // TODO (#2106): replace this example play with an actual implementation
 
             // The angle between each robot spaced out in a circle around the ball
             Angle angle_between_robots =
@@ -50,22 +62,11 @@ void PassEndurancePlay::getNextTactics(TacticCoroutine::push_type &yield,
                     0);
             }
         }
-        else
-        {
-            // line up along center line
-            int initial_offset = static_cast<int>(-move_tactics.size() / 2 + 1);
-            for (size_t k = 0; k < move_tactics.size(); k++)
-            {
-                auto next_position = Point(
-                    world.field().centerPoint().x(),
-                    (initial_offset + static_cast<int>(k)) * 4 * ROBOT_MAX_RADIUS_METERS);
-                move_tactics[k]->updateControlParams(next_position, Angle::zero(), 0);
-            }
-        }
         result.insert(result.end(), move_tactics.begin(), move_tactics.end());
-        yield({{result}});
+        yield({result});
     } while (true);
 }
 
 // Register this play in the genericFactory
-static TGenericFactory<std::string, Play, PassEndurancePlay, PlayConfig> factory;
+static TGenericFactory<std::string, Play, ScoringWithStaticDefendersPlay, PlayConfig>
+    factory;
