+++ src/software/ai/hl/stp/play/hardware_challenge_plays/scoring_with_static_defenders_play.h
